+++ gen/com/yanncebron/m68kplugin/lang/M68kExpressionParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Authors
+ * Copyright 2025 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,13 +30,14 @@ public class M68kExpressionParser {
   static boolean bracket_paren_expression(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "bracket_paren_expression")) return false;
     if (!nextTokenIsFast(b, L_BRACKET)) return false;
-    boolean r;
-    Marker m = enter_section_(b);
+    boolean r, p;
+    Marker m = enter_section_(b, l, _NONE_);
     r = consumeTokenFast(b, L_BRACKET);
-    r = r && expression(b, l + 1, -1);
-    r = r && consumeToken(b, R_BRACKET);
-    exit_section_(b, m, null, r);
-    return r;
+    p = r; // pin = 1
+    r = r && report_error_(b, expression(b, l + 1, -1));
+    r = p && consumeToken(b, R_BRACKET) && r;
+    exit_section_(b, l, m, r, p, null);
+    return r || p;
   }
 
   /* ********************************************************** */
@@ -44,13 +45,14 @@ public class M68kExpressionParser {
   static boolean plain_paren_expression(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "plain_paren_expression")) return false;
     if (!nextTokenIsFast(b, L_PAREN)) return false;
-    boolean r;
-    Marker m = enter_section_(b);
+    boolean r, p;
+    Marker m = enter_section_(b, l, _NONE_);
     r = consumeTokenFast(b, L_PAREN);
-    r = r && expression(b, l + 1, -1);
-    r = r && consumeToken(b, R_PAREN);
-    exit_section_(b, m, null, r);
-    return r;
+    p = r; // pin = 1
+    r = r && report_error_(b, expression(b, l + 1, -1));
+    r = p && consumeToken(b, R_PAREN) && r;
+    exit_section_(b, l, m, r, p, null);
+    return r || p;
   }
 
   /* ********************************************************** */