+++ tests/com/yanncebron/m68kplugin/lang/resolve/M68kLabelResolveTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Authors
+ * Copyright 2025 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -136,6 +136,11 @@ public class M68kLabelResolveTest extends BasePlatformTestCase {
     assertTrue(builtinSymbolPresentation.isItemTextItalic());
     assertEquals(" Current PC", builtinSymbolPresentation.getTailText());
     assertEquals("any", builtinSymbolPresentation.getTypeText());
+  }
+
+  public void testCompletionVariantsInParenthesizedExpression() {
+    myFixture.testCompletionVariants("completionVariantsInParenthesizedExpression.s",
+      getCompletionVariants("label"));
   }
 
   public void testResolveLocalLabelInCorrectScope() {